Allow deletion with GCS (new --single-delete flag)
The GCS S3 compatibility API doesn't support a bulk delete.  This patch
provides a new flag, --single-delete, which allows Warp deletion to work
against GCS.  The flag makes all object deletions be single-object
requests.  Since this is a niche workaround for a missing S3
compatibility feature, the new flag is hidden.

Without specifing the new flag, deletion behavior is unchanged.

NOTE: detecting when GCS is being targeted is outside the scope of a
benchmark tool, so that is not included.